requirements.txt
datacollective
python-dotenv

# Notebook
jupyter
ipython

# Audio processing & visualization
librosa
matplotlib
pandas
numpy
